make general utils and move to using it
